requirements.txt
requests
beautifulsoup4
pyyaml
schedule
